chore: moving Avatar to Molecules
