+++ src/Molecules/Avatar/Avatar.types.ts
